UI interaction for "form" save
Not really using form semantics, as I don’t know how to do this well on
Elm.

We update the model upon every change to any of the fields. When the
user hits ‘save’, the current model is used to populate the database.
Looks like this is the recommended approach.